@@ -1,16 +1,31 @@
-var path = require('path');
-var config = {
-  entry: path.resolve(__dirname, 'src/app.js'),
-  output: {
-    filename: 'bundle.js'
-  },
+const path = require('path');
+ 
+module.exports = {
+  mode: "development",
+  entry: path.resolve(__dirname, './src/index.js'),
   module: {
-    loaders: [{
-      test: /\.jsx?$/,
-      loader: 'babel',
-      include: path.join(__dirname, 'src')
-    }]
-  }
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: ['babel-loader'],
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['*', '.js', '.jsx'],
+  },
+  output: {
+    path: path.resolve(__dirname, './'),
+    filename: 'bundle.js',
+  },
+  devServer: {
+    static: path.join(__dirname, 'public'), // boolean | string | array, static file location
+    // compress: true, // enable gzip compression
+    // historyApiFallback: true, // true for index.html upon 404, object for multiple paths
+    // hot: true, // hot module replacement. Depends on HotModuleReplacementPlugin
+    // https: false, // true for self-signed, object for cert authority
+    // // noInfo: true, // only errors & warns on hot reload
+    // ...
+  },
 };
-
-module.exports = config;
